requirements.txt
# This file may be used to create an environment using:
# $ pip install -r <this file>
# platform: linux-64

zarr
ray
numpy
torch
matplotlib
scikit-image
scipy
pandas
pyarrow
fastparquet
pyyaml
timm

# _libgcc_mutex=0.1=main
# _openmp_mutex=5.1=1_gnu
# aiobotocore=2.7.0=pypi_0
# aiohttp=3.8.6=pypi_0
# aiohttp-cors=0.7.0=pypi_0
# aioitertools=0.11.0=pypi_0
# aiosignal=1.3.1=pypi_0
# asciitree=0.3.3=pypi_0
# asttokens=2.4.1=pypi_0
# async-timeout=4.0.3=pypi_0
# attrs=23.1.0=pypi_0
# blessed=1.20.0=pypi_0
# botocore=1.31.64=pypi_0
# bzip2=1.0.8=h7b6447c_0
# ca-certificates=2023.08.22=h06a4308_0
# cachetools=5.3.2=pypi_0
# certifi=2023.7.22=pypi_0
# charset-normalizer=3.3.2=pypi_0
# click=8.1.7=pypi_0
# colorful=0.5.5=pypi_0
# comm=0.1.4=pypi_0
# contourpy=1.1.1=pypi_0
# cramjam=2.7.0=pypi_0
# cycler=0.12.1=pypi_0
# debugpy=1.8.0=pypi_0
# decorator=5.1.1=pypi_0
# distlib=0.3.7=pypi_0
# executing=2.0.1=pypi_0
# fasteners=0.19=pypi_0
# fastparquet=2023.10.1=pypi_0
# filelock=3.13.1=pypi_0
# fonttools=4.43.1=pypi_0
# frozenlist=1.4.0=pypi_0
# fsspec=2023.10.0=pypi_0
# google-api-core=2.12.0=pypi_0
# google-auth=2.23.4=pypi_0
# googleapis-common-protos=1.61.0=pypi_0
# gpustat=1.1.1=pypi_0
# grpcio=1.59.2=pypi_0
# idna=3.4=pypi_0
# imageio=2.31.6=pypi_0
# ipykernel=6.26.0=pypi_0
# ipython=8.17.2=pypi_0
# jedi=0.19.1=pypi_0
# jinja2=3.1.2=pypi_0
# jmespath=1.0.1=pypi_0
# jsonschema=4.19.2=pypi_0
# jsonschema-specifications=2023.7.1=pypi_0
# jupyter-client=8.5.0=pypi_0
# jupyter-core=5.5.0=pypi_0
# kiwisolver=1.4.5=pypi_0
# lazy-loader=0.3=pypi_0
# ld_impl_linux-64=2.38=h1181459_1
# libffi=3.4.4=h6a678d5_0
# libgcc-ng=11.2.0=h1234567_1
# libgomp=11.2.0=h1234567_1
# libstdcxx-ng=11.2.0=h1234567_1
# libuuid=1.41.5=h5eee18b_0
# markupsafe=2.1.3=pypi_0
# matplotlib=3.8.1=pypi_0
# matplotlib-inline=0.1.6=pypi_0
# mpmath=1.3.0=pypi_0
# msgpack=1.0.7=pypi_0
# multidict=6.0.4=pypi_0
# ncurses=6.4=h6a678d5_0
# nest-asyncio=1.5.8=pypi_0
# networkx=3.2.1=pypi_0
# numcodecs=0.12.1=pypi_0
# numpy=1.26.1=pypi_0
# nvidia-cublas-cu12=12.1.3.1=pypi_0
# nvidia-cuda-cupti-cu12=12.1.105=pypi_0
# nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
# nvidia-cuda-runtime-cu12=12.1.105=pypi_0
# nvidia-cudnn-cu12=8.9.2.26=pypi_0
# nvidia-cufft-cu12=11.0.2.54=pypi_0
# nvidia-curand-cu12=10.3.2.106=pypi_0
# nvidia-cusolver-cu12=11.4.5.107=pypi_0
# nvidia-cusparse-cu12=12.1.0.106=pypi_0
# nvidia-ml-py=12.535.108=pypi_0
# nvidia-nccl-cu12=2.18.1=pypi_0
# nvidia-nvjitlink-cu12=12.3.52=pypi_0
# nvidia-nvtx-cu12=12.1.105=pypi_0
# opencensus=0.11.3=pypi_0
# opencensus-context=0.1.3=pypi_0
# openssl=3.0.11=h7f8727e_2
# packaging=23.2=pypi_0
# pandas=2.1.2=pypi_0
# parso=0.8.3=pypi_0
# pexpect=4.8.0=pypi_0
# pillow=10.0.1=pypi_0
# pip=23.3=py311h06a4308_0
# platformdirs=3.11.0=pypi_0
# prometheus-client=0.18.0=pypi_0
# prompt-toolkit=3.0.39=pypi_0
# protobuf=4.24.4=pypi_0
# psutil=5.9.6=pypi_0
# ptyprocess=0.7.0=pypi_0
# pure-eval=0.2.2=pypi_0
# py-spy=0.3.14=pypi_0
# pyarrow=14.0.0=pypi_0
# pyasn1=0.5.0=pypi_0
# pyasn1-modules=0.3.0=pypi_0
# pydantic=1.10.13=pypi_0
# pygments=2.16.1=pypi_0
# pyparsing=3.1.1=pypi_0
# python=3.11.5=h955ad1f_0
# python-dateutil=2.8.2=pypi_0
# pytz=2023.3.post1=pypi_0
# pyyaml=6.0.1=pypi_0
# pyzmq=25.1.1=pypi_0
# ray=2.7.1=pypi_0
# readline=8.2=h5eee18b_0
# referencing=0.30.2=pypi_0
# requests=2.31.0=pypi_0
# rpds-py=0.10.6=pypi_0
# rsa=4.9=pypi_0
# s3fs=2023.10.0=pypi_0
# sbrnet-core=0.1=dev_0
# scikit-image=0.22.0=pypi_0
# scipy=1.11.3=pypi_0
# setuptools=68.0.0=py311h06a4308_0
# six=1.16.0=pypi_0
# smart-open=6.4.0=pypi_0
# sqlite=3.41.2=h5eee18b_0
# stack-data=0.6.3=pypi_0
# sympy=1.12=pypi_0
# tifffile=2023.9.26=pypi_0
# tk=8.6.12=h1ccaba5_0
# torch=2.1.0=pypi_0
# torchvision=0.16.0=pypi_0
# tornado=6.3.3=pypi_0
# traitlets=5.13.0=pypi_0
# triton=2.1.0=pypi_0
# typing-extensions=4.8.0=pypi_0
# tzdata=2023.3=pypi_0
# urllib3=2.0.7=pypi_0
# virtualenv=20.21.0=pypi_0
# wcwidth=0.2.9=pypi_0
# wheel=0.41.2=py311h06a4308_0
# wrapt=1.15.0=pypi_0
# xz=5.4.2=h5eee18b_0
# yarl=1.9.2=pypi_0
# zarr=2.16.1=pypi_0
# zlib=1.2.13=h5eee18b_0